* co_snprintf(), co_debug...(), co_terminal_print...(), co_ntevent_print():   Add '__attribute__' for printf format checking at compile time.   Adjust %x into %lx or %p, %d into %ld. Cast for self typedefed int`s. * GetLastError(), NTSTATUS: Print always as hex.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/devel@708 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/os/winnt/kernel/lowlevel/time.c
@@ -62,7 +62,7 @@ unsigned long co_os_get_cpu_khz(void)
 		    		query, NULL, NULL);
 
 		if (status != STATUS_SUCCESS)
-			co_debug("getting MHz failed %x", status);
+			co_debug("getting MHz failed %x", (int)status);
 	}
 
 	return(1000*mhz);
